Move TenantSettings to Tenant , update app version
 - Updated app version.
