+++ TLabs.ExchangeSdk/Farming/Dtos/TenantCreateDto.cs
@@ -9,4 +9,6 @@ public class TenantCreateDto
     public string TokenLogoIconId { get; set; }
     public string TokenName { get; set; }
     public string Url { get; set; }
+    public bool IsShowMiniGame { get; set; }
+    public string ColorSettingsJson { get; set; }
 }
