+++ TLabs.ExchangeSdk/Farming/Dtos/TenantSettingsDto.cs
@@ -2,8 +2,8 @@ using System.Collections.Generic;
 
 namespace TLabs.ExchangeSdk.Farming.Dtos;
 
-public class TenantSettingsCreateDto
+public class TenantSettingsDto
 {
     public bool IsShowMiniGame { get; set; }
-    public Dictionary<string, string> ColorSettings { get; set; }
+    public string ColorSettingsJson { get; set; }
 }
