+++ TLabs.ExchangeSdk/Farming/Tenant.cs
@@ -1,4 +1,6 @@
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Text.Json;
 using System.Text.Json.Serialization;
 
 namespace TLabs.ExchangeSdk.Farming;
@@ -27,7 +29,33 @@ public class Tenant
 
     public string Url { get; set; }
 
-    public long TenantSettingsId { get; set; }
+    public bool IsShowMiniGame { get; set; }
 
-    public TenantSettings TenantSettings { get; set; }
+    public string ColorSettingsJson { get; private set; }
+
+    public Dictionary<string, string> GetColorSettings()
+    {
+        if (string.IsNullOrEmpty(ColorSettingsJson))
+            return new Dictionary<string, string>();
+
+        return JsonSerializer.Deserialize<Dictionary<string, string>>(ColorSettingsJson);
+    }
+
+    public void SetColorSettings(Dictionary<string, string> colorSettings)
+    {
+        ColorSettingsJson = JsonSerializer.Serialize(colorSettings);
+    }
+
+    public string GetColor(string colorName)
+    {
+        var settings = GetColorSettings();
+        return settings.GetValueOrDefault(colorName);
+    }
+
+    public void SetColor(string colorName, string colorValue)
+    {
+        var settings = GetColorSettings();
+        settings[colorName] = colorValue;
+        SetColorSettings(settings);
+    }
 }
