+++ TLabs.ExchangeSdk/TLabs.ExchangeSdk.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
-    <Version>1.0.437</Version>
+    <Version>1.0.438</Version>
   </PropertyGroup>
 
   <PropertyGroup>
